requirements.txt
# 基本依賴
pyvisa==1.13.0
pyvisa-py==0.7.1
numpy==1.24.3
matplotlib==3.7.1

# 串口通訊 (Rigol DP711)
pyserial==3.5

# 數據處理
pandas==2.0.3

# 科學計算
scipy==1.11.1

# GUI 介面
PyQt6==6.5.2
pyqtgraph==0.13.3

# 注意：socket 和 datetime 是Python內建模組，無需安裝